Extend JQL to only match issue url from header
Previously the bot will search for the issue URL in the whole desc of
existing tasks. This commit updates it to only compare the url with the
first sentence of the tasks, since this was added by the bot.

This will avoid not opening issues, if their URL was referenced by
another issue.